refactor(SPV-1173): fix import paths.
