+++ examples/fetch_access_keys/fetch_access_keys.go
@@ -5,8 +5,8 @@ import (
 	"log"
 
 	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples/exampleutil"
 )
 
 func main() {
